refactor: Update Flutter 3.35.4 compatibility and clean up legacy assets - Remove outdated Chad image assets - Update generated plugin registrants - Sync platform configuration files - Clean up legacy documentation files

+++ lib/screens/home/widgets/progress_card_widget.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import '../../../generated/app_localizations.dart';
 import '../../../utils/constants.dart';
 
 /// 프로그램 진행률을 표시하는 카드 위젯
@@ -112,13 +111,13 @@ class ProgressCardWidget extends StatelessWidget {
   }
 
   Widget _buildProgressContent(BuildContext context, ThemeData theme) {
-    final currentWeek = programProgress?.weeklyProgress.currentWeek ?? 1;
-    final totalWeeks = programProgress?.totalWeeks ?? 6;
-    final progressPercentage = programProgress?.progressPercentage ?? 0.0;
-    final completedDaysThisWeek = programProgress?.completedDaysThisWeek ?? 0;
-    final totalDaysThisWeek = programProgress?.totalDaysThisWeek ?? 3;
-    final completedSessions = programProgress?.completedSessions ?? 0;
-    final totalSessions = programProgress?.totalSessions ?? 18;
+    final currentWeek = (programProgress?.weeklyProgress.currentWeek ?? 1) as int;
+    final totalWeeks = (programProgress?.totalWeeks ?? 6) as int;
+    final progressPercentage = (programProgress?.progressPercentage ?? 0.0) as double;
+    final completedDaysThisWeek = (programProgress?.completedDaysThisWeek ?? 0) as int;
+    final totalDaysThisWeek = (programProgress?.totalDaysThisWeek ?? 3) as int;
+    final completedSessions = (programProgress?.completedSessions ?? 0) as int;
+    final totalSessions = (programProgress?.totalSessions ?? 18) as int;
 
     return Container(
       padding: const EdgeInsets.all(16),
